requirements.txt
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
statsmodels>=0.14.0
scikit-learn>=1.4.0
joblib>=1.3.0
Flask>=3.0.0
gunicorn>=21.2.0
python-dateutil>=2.8.2
prophet>=1.1.5
flask
gunicorn